fixed alignment issue between the nav bar and the tables
http://stackoverflow.com/questions/27827315/empty-white-space-above-uita
bleview-inside-a-uiview?answertab=votes#tab-top

Try to look in the 'attribute inspector' (in the right menu) of the
Participants ViewController. Check for the option 'Extend Edges' and
uncheck the 'Under Top Bars', and then relocate your tableview.

Another possible solution is to just uncheck the option 'Adjust Scroll
View Insets'. The reason is that when the viewController extends its
edges, let's say under the top bar, the tableView's scrollView
automatically adjusts its inset top, so that the content of the
tableView will start exactly under the top bar. But in your case this
is not needed, since your tableView itself starts under the bar.